requirements.txt
Flask==2.3.1
Frozen-Flask==0.18
pandas==2.0.3
Werkzeug==2.3.1
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.23.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0